% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleNiObject.R
\encoding{UTF-8}
\name{assembleNiObject}
\alias{assembleNiObject}
\title{Assemble Data Set}
\usage{
assembleNiObject(
  inputData = NULL,
  indicators = NULL,
  referenceValues = NULL,
  indicatorObservations = NULL,
  ICunits = NULL,
  BSunits = NULL,
  ecosystems = NULL,
  NIunits = NULL,
  predefNIunits = c(allArea = T, parts = F, counties = F),
  indexType = "ecosystem",
  part = "ecosystem",
  total = "total",
  partOfTotal = 0
)
}
\arguments{
\item{inputData}{list with elements named
\tabular{ll}{\code{[[1]]} \tab \code{"indicators"},\cr
             \code{[[2]]} \tab \code{"referenceValues"},\cr
             \code{[[3]]} \tab \code{"indicatorObservations"},\cr
             \code{[[4]]} \tab \code{"ICunits"},\cr
             \code{[[5]]} \tab \code{"BSunits"},\cr
             \code{[[6]]} \tab \code{"ecosystems"},\cr
             \code{[[7]]} \tab \code{"NIunits"},}
where the second and third elements are lists, while the others are data.frames.
The last two element, \code{ecosystems} and \code{NIunits}, may be
omitted or set to \code{NULL}. For example lists of class \code{niDataImport}.}

\item{indicators}{data.frame containing indicator characteristics.}

\item{referenceValues}{list containing reference value characteristics.}

\item{indicatorObservations}{list containing indicator observations characteristics.}

\item{ICunits}{data.frame containing ICunits characteristics.}

\item{BSunits}{data.frame containing BSunits characteristics.}

\item{ecosystems}{data.frame containing ecosystems characteristics.}

\item{NIunits}{data.frame containing characteristics of user defined NIunits.}

\item{predefNIunits}{named logical vector defining NIunits to be associated
with the data set. Default settings are
\code{c(allArea = T, parts = F, counties = F)}}

\item{indexType}{character. Valid \code{indexTypes} are
\code{"ecosystem"} (default), \code{"thematic"}, \code{"thematic per ecosystem"}.}

\item{part}{character, one of \code{c("ecosystem","marine","terrestrial")}.
Numerator variable in criterion for selecting BSunits.
\cr\code{"ecosystem"}: area of major ecosystem in BSunit.
\cr\code{"marine"}: marine area \cr\code{"terrestrial"}: terrestrial area.}

\item{total}{character, one of \code{c("marine","terrestrial","total")}.
Denominator variable in criterion for selecting BSunits.
\cr\code{"total"}: total area of BSunit.}

\item{partOfTotal}{numerical scalar in \code{[0,1]}}
}
\value{
A list of class \code{\link{niInput}} with elements
  \tabular{lll}{\code{[[1]]} \tab \code{indicators} \tab data.frame with indicator data.\cr
  \code{[[2]]} \tab \code{ICunits} \tab integer BSunit x indicator matrix of ICunits.\cr
  \code{[[3]]} \tab \code{BSunits} \tab data.frame of BSunit data.\cr
  \code{[[4]]} \tab \code{referenceValues} \tab data.frame of reference values.\cr
  \code{[[5]]} \tab \code{indicatorValues} \tab list of data.frames of indicator values.\cr
  \code{[[6]]} \tab \code{NIunits} \tab 0/1 BSunit x NIunit matrix of NIunits.}
}
\description{
Assembles and structures data into a complete and consistent data set
for calculating the Nature Index or thematic indices.
}
\details{
\code{assembleNiObject} assembles and restructures a data set into a list of
class \code{\link{niInput}} or a list of such lists. Most of the data are
typically imported from the Nature Index database by use of
\code{\link{importDatasetApi}}, except the delineation of NIunits (see below).
\code{niInput} objects have the particular structure and format required for
data input to the function \code{\link{calculateIndex}}.

NIunits are geographical areas for which the indices are to be calculated.
NIunits may be defined in two ways, either by choosing from predefined units
(argument \code{predefNIunits}), and/or entered as a data.frame with 'user
defined' units (argument \code{NIunits}). The data.frame should contain
variables named \code{NIunitName} and \code{BSunitId} naming NIunits and
defining them in terms of BSunits.

The data set to be assembled may be entered either as a list assigned to the
argument \code{inputData} containing all data elements (see \code{Arguments} above),
or as several objects, one for each element, assigned to other function arguments.

The structure of the assembled data set depends on what type of index to be
calculated (argument \code{indexType}): the Nature Index for major
ecosystem(s) (\code{indexType = "ecosystem"}, default), a thematic index
(\code{indexType = "thematic"}), or a thematic index calculated for a set of
major ecosystems (\code{indexType = "thematic per ecosystem"}).

Parameters \code{part}, \code{total}, and \code{partOfTotal} may be used
to restrict the data set to BSunits satisfying: \code{part/total > partOfTotal},
where \code{part} and \code{total} are names of variables in the
\code{BSunits} data.frame containing area data.

\code{assembleNiObject} calls \code{\link{checkInputData}} to check if the
data set contains all necessary information for calculating the Nature Index,
and whether the data are consistent.

\code{assembleNiObject} also calls \code{\link{elicitate}} to elicitate all
observations in the data set that are not given as custom distributions.
}
\examples{
\dontrun{
carnivoreImport <- importDatasetApi(
     username = "...",
     password = "...",
     indic = c("Jerv","Gaupe","Ulv"),
     year = c("1990","2000","2010","2014","2019"))
carnivoreInput <- assembleNiObject(inputData = carnivoreImport,
                    predefNIunits = c(allArea = T, parts = T, counties = F),
                    indexType = "thematic")

mountainImport <- importDatasetApi(
     username = "...",
     password = "...",
     eco = "Fjell",
     year = c("1990","2000","2010","2014","2019"))
mountainInput <- assembleNiObject(inputData = mountainImport,
                    predefNIunits = c(allArea = T, parts = T, counties = F),
                    indexType = "ecosystem",
                    part = "ecosystem",
                    total = "terrestrial",
                    partOfTotal = 0.2))
}

}
\seealso{
\code{\link{checkInputData}}, \code{\link{elicitate}},
and \code{\link{calculateIndex}}.
\cr The vignette \code{objectsInNIcalc} gives a more detailed description
of \code{\link{niDataImport}} and \code{\link{niInput}} lists.
}
\author{
Bård Pedersen
}
