% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BSunits}
\alias{BSunits}
\title{BSunits}
\format{
A data frame with 4 rows and 5 variables:
\describe{
  \item{\code{BasicunitID}}{integer COLUMN_DESCRIPTION}
  \item{\code{Basicunit}}{character COLUMN_DESCRIPTION}
  \item{\code{Area}}{double COLUMN_DESCRIPTION}
  \item{\code{NIarea1}}{integer COLUMN_DESCRIPTION}
  \item{\code{NIarea2}}{integer COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
BSunits
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
