% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIndex.R
\encoding{UTF-8}
\name{calculateIndex}
\alias{calculateIndex}
\title{Calculate Index}
\usage{
calculateIndex(x = NULL, imputations = NULL, awBSunit = NULL, nsim = 1000, ...)
}
\arguments{
\item{x}{list of class \code{niInput}.}

\item{imputations}{optional list of class \code{niImputations}.}

\item{awBSunit}{character, name of variable in \code{x$BSunit} used to calculate
NIunit weights.}

\item{nsim}{integer - number of bootstrap simulations, default is 1000.}

\item{...}{further arguments passed on to \code{\link{calculateBSunitWeights}}
(\code{fids, tgroups, keys, w}), \code{\link{calculateNIunitWeights}}
(\code{awbs, awBSunit}), and \code{\link{scaleObsMat}}
(\code{truncAtRef}).}
}
\value{
\code{calculateIndex} returns a list of class \code{niOutput} where each
list element is itself a list of class \code{niSeries} containing the index results
for an NIunit. Each element of an \code{niSeries} list again is a list of
class \code{niValue}. Objects of class \code{niValue} contain the results from
the calculation of a single Nature Index or thematic index value.
}
\description{
Calculate the Nature Index for a major habitat, or calculate a thematic
index.
}
\details{
A time series of index values are calculated for each of one or more NIunits.
NIunits as well as years included in the time series are specified in
\code{x}.

Index calculations and associated terminology are explained in the vignette
\code{NatureIndexCalculation}.

\code{calculateIndex} produces an extensive output for each index value
to facilitate further analyses of the results. In addition to a sample of
\code{nsim} draws of the index estimate, output includes data describing
indicators, ICunits, and BSunits included in the input data set, imputations
if present in the data, weights per combination of BSunits and indicators,
calculation parameters etc.

The vignette \code{objectsInNIcalc} gives a detailed description of output
objects generated by \code{calculateIndex} and provides an overview of
methods developed for \code{niOutput}, \code{niSeries}, and \code{niValue}
classes.
}
\examples{
\dontrun{
# Calculation of a thematic index:

themeImputes <- imputeData(x = themeData,
                               nSim = 1000,
                               transConst = 0.01,
                               maxit = 20,
                               printFlag = TRUE)
carnivoreIndex <- calculateIndex(x = themeData,
                                 imputations = themeImputes,
                                 nsim = 1000,
                                 fids = FALSE,
                                 tgroups = FALSE,
                                 keys = "ignore",
                                 w = 0,
                                 awbs = TRUE,
                                 awBSunit = "Skog")
}

}
\seealso{
\code{\link{niInput}} and \code{\link{niImputations}} for more
detailed description of input data, \code{\link{niOutput}} for more detailed
description of class \code{niOutput} objects. \cr The following methods are
available within \code{NIcalc} to display and summarize results produced by
\code{calculateIndex}: \code{\link{summary.niOutput}},
\code{\link{summary.niSeries}}, \code{\link{plot.niSeries}},
\code{\link{summaryWeights}}, \code{\link{plot.niValue}}, and
\code{\link{plotWeights}}. \cr \code{calculateIndex} calls a series of
other functions within NIcalc: \code{\link{calculateBSunitWeights}},
\code{\link{calculateNIunitWeights}}, \code{\link{calculateWeights}},
\code{\link{sampleObsMat}},\code{\link{scaleObsMat}},
\code{\link{indexCalculation}}, and \code{\link{indexCalculationPerBSunit}}.
}
\author{
Bård Pedersen
}
