% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputData.R
\encoding{UTF-8}
\name{checkInputData}
\alias{checkInputData}
\title{Check Data Set}
\usage{
checkInputData(
  inputData = NULL,
  indicators = NULL,
  referenceValues = NULL,
  indicatorObservations = NULL,
  ICunits = NULL,
  BSunits = NULL,
  ecosystems = NULL,
  NIunits = NULL
)
}
\arguments{
\item{inputData}{list with elements named
\tabular{ll}{\code{[1]} \tab \code{"indicators"},\cr
             \code{[2]} \tab \code{"referenceValues"},\cr
             \code{[3]} \tab \code{"indicatorObservations"},\cr
             \code{[4]} \tab \code{"ICunits"},\cr
             \code{[5]} \tab \code{"BSunits"},\cr
             \code{[6]} \tab \code{"ecosystems"},\cr
             \code{[7]} \tab \code{"NIunits"},}
where the second and third elements are lists, while the others are data.frames.
The last two element, \code{"ecosystems"} and \code{"NIunits"}, may be
omitted or set to \code{NULL}.}

\item{indicators}{data.frame containing indicator characteristics.}

\item{referenceValues}{list containing reference value characteristics.}

\item{indicatorObservations}{list containing indicator observations characteristics.}

\item{ICunits}{data.frame containing ICunits characteristics.}

\item{BSunits}{data.frame containing BSunits characteristics.}

\item{ecosystems}{data.frame containing ecosystems characteristics.}

\item{NIunits}{data.frame containing NIunits characteristics.}
}
\value{
either a list with the same elements that was
  assigned to function arguments, \cr\cr or \cr\cr a set of error messages.
}
\description{
Checks the content and consistency of a candidate data set for calculating
the Nature Index.
}
\details{
\code{checkInputData} runs through an extensive series of tests checking
whether the entered dataset contains all the necessary data objects and
variables for calculating the Nature Index, and whether the data contain
consistent information.

The function checks
\itemize{
\item that all necessary types of data objects are included:
      \itemize{\item indicators,
               \item reference values, \item indicator observations,
               \item ICunits, \item BSunits, \item ecosystems (not required),
               \item NIunits (not required)}
\item that all required variables characterizing each object type are included.
\item that IDs and/or names are unique for all types of objects.
\item consistency among input objects of different types.
\item for missing values.
}

While running through the tests, \code{checkInputData} generates error
messages or warnings when identifying missing or inconsistent information.
The function may make some simple modifications to the data set to mend
shortcomings. It issues warning messages explaining such modifications.

The data set to be checked may be entered either as a list assigned to the
argument \code{inputData} and containing all types of objects, or as several
objects, one for each type, assigned to the other function arguments.
}
\examples{
\dontrun{
carnivoreImport <- importDatasetApi(
     username = "...",
     password = "...",
     eco = NULL,
     indic = c("Jerv","Gaupe","Ulv"),
     year = c("1990","2000","2010","2014","2019"))
checkInputData(inputData = carnivoreImport)

checkInputData(BSunits = 1)
}

}
\seealso{
\code{\link{assembleNiObject}} and \code{\link{niInput}} for more
about the requirements to input data,
\code{\link{importDatasetApi}}, \code{\link{getNIdata}}, and
\code{\link{niDataImport}} for importing data from the NI database.
}
\author{
Bård Pedersen
}
