% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimlight.fct.R
\name{estimlight.fct}
\alias{estimlight.fct}
\title{Fit Probability Distribution to one Indicator Observation, reduced Version}
\usage{
estimlight.fct(obsval = NULL, proba = c(0.25, 0.75), type = "continuous")
}
\arguments{
\item{obsval}{double, length = 3, observed mean and quantiles in the sequence
\code{c(lower.quantile, mean, upper.quantile)}}

\item{proba}{double, length = 2, quantiles supplied in \code{"obsval"}.
Default is the lower and upper quartiles}

\item{type}{character, length = 1, type of measurement scales. Valid types
are \code{"continuous"} and \code{"discrete"}.
When \code{type = "continuous"}, a continuous model is fitted to the indicator
observation. When \code{type = "discrete"}, a discrete model is fitted.}
}
\value{
\code{estim.fct} returns a data.frame with \code{dim = c(1,4)},
consisting of the following vectors\cr \code{[[1]] $distrib} character,
selected family for model distribution, i.e. one of \code{c("LogNormal",
"NegBinom", "Poisson")}. \cr \code{[[2]] $mu} double,
first parameter of fitted model distribution\cr \code{[[3]] $sig}
double, second parameter of fitted model distribution\cr
\code{[[4]] $crit} double, sum of squared deviations between observed
parameters and those of the fitted model distribution.
}
\description{
Function that, in the continuous case, fits a lognormal distribution to an
expected value and two quantiles, or fits a Poisson- or negative binomial
distribution in the discrete case, using the least square criterion.
\code{estimlight.fct} is a simplified alternative to \code{\link{estim.fct}}.
}
\examples{
estimlight.fct(obsval = c(6,13,25))
estimlight.fct(obsval = c(6,13,25),proba = c(0.025,0.975), type = "continuous")
estimlight.fct(obsval = c(6,13,25),proba = c(0.025,0.975), type = "discrete")
}
\author{
Nigel Yoccoz and Bård Pedersen
}
\seealso{
\code{\link{estim.fct}},\cr \code{\link{elicitate}} for fitting
probability distributions to multiple indicator observations and for list of
model distributions included,\cr \code{\link{qdev}} for calculating sum of
squares between \code{obsval} and model.
}

