% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndPlotDistribution.R
\name{fitAndPlotDistibution}
\alias{fitAndPlotDistibution}
\title{fitAndPlotDistribution}
\usage{
fitAndPlotDistibution(obsval = c(mu = NULL, lower = NULL, upper = NULL),
  refValue = 1, probQuant = c(lower = 0.25, upper = 0.75), ...)
}
\arguments{
\item{obsval}{named vector containing mu and lower and upper quantiles to fit the distribution to.}

\item{refValue}{the reference value to be drawn in the  plot}

\item{probQuant}{named vector containing the lower and upper quantiles to draw in the plot}
}
\description{
Calls estimate (elicitation) functions with inputs and plots the resulting distribution
}
