% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndicators.R
\encoding{UTF-8}
\name{getIndicators}
\alias{getIndicators}
\title{List the indicators that the user is permitted to alter}
\usage{
getIndicators()
}
\value{
A data frame containing indicator IDs and names.
The indicator IDs are entered as parameters to the function \code{getIndicatorValues}
in order to retrieve the current values for a given indicator.
}
\description{
Based on access to the Nature Index database through a token generated by
\code{getToken}, this function lists the indicators (ID and names) that the user
is responsible for
}
\examples{
\dontrun{
getIndicators()
}

}
\seealso{
\code{\link{getIndicatorValues}} and \code{\link{getToken}}.
}
\author{
Jens Åström
}
