% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndicatorValues.R
\name{getIndicatorValues}
\alias{getIndicatorValues}
\title{Get the present indicator values from the Nature Index database}
\usage{
getIndicatorValues(indicatorID = NULL, token = niToken)
}
\arguments{
\item{indicatorID}{numerical    a numerical value identifying the indicator. Usually this is first retreived by the \link{getIndicators} function}
}
\value{
A list containing 1) a data frame of the values of the indicator and 2) a list of distribution objects identifying the uncertainties of the estimates (optional)
}
\description{
This function retreives the current values for a given indicator. It is used both to check the current
information, and as a starting point for updating the data. In case of updates, individual rows are altered
by the function setIndicatorValue.
}
\section{Notes}{

}
\examples{
\dontrun{
myValues <- getIndicatorValues(IndicatorID = 1)
}

}
\author{
Jens Åström
}

