% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNIdata.R
\encoding{UTF-8}
\name{getNIdata}
\alias{getNIdata}
\alias{getBSunits}
\alias{getEcosystems}
\alias{getEcosystemBSunitData}
\alias{getFunctionalGroupData}
\alias{getICunits}
\alias{getICunitBSunit}
\alias{getIndicators2}
\alias{getIndicatorEcosystemData}
\alias{getIndicatorValues2}
\alias{getRefYearData}
\alias{getScalingModelData}
\title{Read a Table in the NI Database}
\usage{
getBSunits(BSunitIds = NULL)

getEcosystems()

getEcosystemBSunitData()

getFunctionalGroupData()

getICunits(indicatorIds = NULL)

getICunitBSunit(icunits = NULL)

getIndicators2(indicatorIds = NULL)

getIndicatorEcosystemData(ecosystemIds = NULL, indicatorIds = NULL)

getIndicatorValues2(indicatorIds = NULL, refYear = NULL)

getRefYearData()

getScalingModelData()
}
\arguments{
\item{BSunitIds}{integer, optional.
\cr\code{getBSunits}: Data for specified BSunits will be retrieved from the
\code{Kommune} table. The default (\code{NULL}) is to retrieve data for all BSunits.}

\item{indicatorIds}{integer, optional. \cr\code{getIndicators2}: Data for specified indicators
will be retrieved from the \code{Indikator} table. The default (\code{NULL}) is to
retrieve data for all indicators.
\cr\code{getIndicatorEcosystemData}: Fidelity of specified indicators to all
ecosystems or those given in \code{ecosystemIds}, will be retrieved from the
\code{Indikator_Okosystem} table.
\cr\code{getICunits}: ICunits for specified indicators will be retrieved from
\code{Omraade}. The default (\code{NULL}) is to retrieve all ICunits.
\cr\code{getIndicatorValues2}: Observations for specified indicators will be
retrieved from \code{Verdier}. The default (\code{NULL}) \emph{is to retrieve no data}.}

\item{icunits}{integer, optional.
\cr\code{getICunitBSunit}: Data for specified ICunits will be retrieved from the
\code{Omraade_Kommune} table. The default (\code{NULL}) is to retrieve BSunits for all ICunits.}

\item{ecosystemIds}{integer, optional.
\cr\code{getIndicatorEcosystemData}: Fidelity of all indicators (or of
those specified in \code{indicatorIds}) to the specified ecosystems, will
be retrieved from the \code{Indikator_Okosystem} table. The default
(\code{ecosystemIds = NULL}) is to retrieve fidelity of all indicators
to all ecosystems.}

\item{refYear}{character, optional.
\cr\code{getIndicatorValues2}: Years of the observations to retrieve and
whether to retreive reference values. E.g. \code{refYear = c("2010","Referanseverdi")}
returns observations from 2010 together with reference values. Default = (\code{NULL}),
which means you get all observations as well as reference values.}
}
\value{
All functions except \code{getIndicatorValues2} returns a data frame
  with content according to the above table, or an error message if extraction of data
  from the database fails.
  \cr\cr\code{getIndicatorValues2} returns a list of two elements:
  \cr 1) a data frame \code{indicatorValues} of indicator observations content
  according to the above table, and
  \cr 2) a list \code{customDistributions} of optional distribution objects associated
  with some of the observations. Distribution objects are identified through
  the variable \code{indicatorValues$customDistributionUUID}
}
\description{
Functions that reads data from tables in the NI database via
[The Nature Index Application Programming Interface](https://www8.nina.no/NaturindeksNiCalc/index.html)
and imports the content as a data.frame.
}
\details{
Each function reads one table in the database:
\tabular{llll}{
Function   \tab Table in NI database   \tab Content
     \tab Column names in returned dataframe\cr
\code{getBSunits} \tab \code{Kommune}
     \tab Characteristics for all BSunits.
     \tab \code{[,1]} \code{"id"},\cr
\tab \tab \tab \code{[,2]} \code{"name"},\cr
\tab \tab \tab \code{[,3]} \code{"area"},\cr
\tab \tab \tab \code{[,4]} \code{"terrestrial"},\cr
\tab \tab \tab \code{[,5]} \code{"marine"},\cr
\tab \tab \tab \code{[,6]} \code{"county"},\cr
\tab \tab \tab \code{[,7]} \code{"part"},\cr
\tab \tab \tab \code{[,8]} \code{"centerLat"},\cr
\tab \tab \tab \code{[,9]} \code{"centerLong"}\cr\cr
\code{getEcosystems} \tab \code{T_Okosystem}
     \tab Ecosystem names and ids.\tab \code{[,1]} \code{"id"},\cr
\tab \tab \tab \code{[,2]} \code{"name"},\cr
\tab \tab \tab \code{[,3]} \code{"nameEnglish"}.\cr\cr
\code{getEcosystemBSunitData} \tab \code{Kommune_Okosystem}
     \tab Area of major ecosystems
     \tab \code{[,1]} \code{"ecosystemId"},\cr
\tab \tab per BSunit \tab \code{[,2]} \code{"municipalityId"},\cr
\tab \tab \tab \code{[,3]} \code{"area"}.\cr\cr
\code{getFunctionalGroupData} \tab \code{T_FunksjonellGruppe}
     \tab names of functional groups
     \tab \code{[,1]} \code{"id"},\cr
\tab \tab \tab \code{[,2]} \code{"name"},\cr
\tab \tab \tab \code{[,3]} \code{"nameEnglish"}.\cr\cr
\code{getICunits} \tab \code{Omraade} \tab ICunit for each indicator
     \tab \code{[,1]} \code{"id"},\cr
\tab \tab \tab \code{[,2]} \code{"name"},\cr
\tab \tab \tab \code{[,3]} \code{"indicatorId"}.\cr\cr
\code{getICunitBSunit} \tab \code{Omraade_Kommune} \tab BSunits within each ICunit.
     \tab \code{[,1]} \code{"areaId"},\cr
\tab \tab \tab \code{[,2]} \code{"municipalityId"}.\cr\cr
\code{getIndicators2} \tab \code{Indikator} \tab Characteristics for all indicators
     \tab \code{[, 1]} \code{"id"},\cr
\tab \tab \tab \code{[, 2]} \code{"active"},\cr
\tab \tab \tab \code{[, 3]} \code{"name"},\cr
\tab \tab \tab \code{[, 4]} \code{"nameEnglish"},\cr
\tab \tab \tab \code{[, 5]} \code{"nameLatin"},\cr
\tab \tab \tab \code{[, 6]} \code{"indicatorTypeId"},\cr
\tab \tab \tab \code{[, 7]} \code{"taxaDetailId"},\cr
\tab \tab \tab \code{[, 8]} \code{"functionalGroupId"},\cr
\tab \tab \tab \code{[, 9]} \code{"scalingModel"},\cr
\tab \tab \tab \code{[,10]} \code{"taxaKingdomId"},\cr
\tab \tab \tab \code{[,11]} \code{"keyElement"},\cr
\tab \tab \tab \code{[,12]} \code{"referenceStateId"}.\cr\cr
\code{getIndicatorEcosystemData} \tab \code{Indikator_Okosystem}
     \tab Fidelity of indicators
     \tab \code{[,1]} \code{"indicatorId"},\cr
\tab \tab per major ecosystem \tab \code{[,2]} \code{"ecosystemId"},\cr
\tab \tab \tab \code{[,3]} \code{"percent"}.\cr\cr
\code{getIndicatorValues2} \tab \code{Verdier}
     \tab Reference values and
     \tab \code{[, 1]} \code{"indicatorId"},\cr
\tab \tab indicator observations \tab \code{[, 2]} \code{"indicatorName"},\cr
\tab \tab \tab \code{[, 3]} \code{"areaId"},\cr
\tab \tab \tab \code{[, 4]} \code{"areaName"},\cr
\tab \tab \tab \code{[, 5]} \code{"yearId"},\cr
\tab \tab \tab \code{[, 6]} \code{"yearName"},\cr
\tab \tab \tab \code{[, 7]} \code{"verdi"},\cr
\tab \tab \tab \code{[, 8]} \code{"nedre_Kvartil"},\cr
\tab \tab \tab \code{[, 9]} \code{"ovre_Kvartil"},\cr
\tab \tab \tab \code{[,10]} \code{"datatypeId"},\cr
\tab \tab \tab \code{[,11]} \code{"datatypeName"},\cr
\tab \tab \tab \code{[,12]} \code{"unitOfMeasurement"},\cr
\tab \tab \tab \code{[,13]} \code{"customDistributionUUID"},\cr
\tab \tab \tab \code{[,14]} \code{"distributionName"},\cr
\tab \tab \tab \code{[,15]} \code{"distributionId"},\cr
\tab \tab \tab \code{[,16]} \code{"distParam1"},\cr
\tab \tab \tab \code{[,17]} \code{"distParam2"}.\cr\cr
\code{getRefYearData} \tab \code{T_RefAar}
     \tab Ids for observation years and
     \tab \code{[,1]} \code{"id"},\cr
\tab \tab the reference value category\tab \code{[,2]} \code{"name"},\cr
\tab \tab \tab \code{[,3]} \code{"nameEnglish"}.\cr\cr
\code{getScalingModelData} \tab \code{T_Skaleringsmodell}
     \tab names of scaling models
     \tab \code{[,1]} \code{"id"},\cr
\tab \tab \tab \code{[,2]} \code{"name"}.
}
}
\section{Note}{

Column names of returned objects are generated by the API and differ somewhat from the
column headings in the corresponding table in the NI database.
}

\examples{
\dontrun{
getBSunits()
getEcosystems()
getEcosystemBSunitData()
getFunctionalGroupData()
getICunits()
getICunitBSunit()
getIndicators2()
getIndicatorEcosystemData()
getIndicatorValues2(indicatorIds = 1:2,refYear = c("2010","Referanseverdi"))
getRefYearData()
getScalingModelData()
}

}
\seealso{
\code{\link{getToken}} for retrieving tokens from the Nature Index API,
 \code{\link{extractContentFromNiapi}}, a general purpose utility function called
 by the functions described here that retrieves data from the specified tables
 in the Nature Index database, and \code{\link{importDatasetApi}} for importing
 from the database all required data for calculating the Nature Index for an
 ecosystem.
}
\author{
Bård Pedersen
}
