% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDatasetApi.R
\encoding{UTF-8}
\name{importDatasetApi}
\alias{importDatasetApi}
\title{Import Dataset}
\usage{
importDatasetApi(
  username = NULL,
  password = NULL,
  eco = NULL,
  indic = NULL,
  year = NULL,
  norwegian = TRUE,
  refYearCode = 0
)
}
\arguments{
\item{username}{character. Your username for accessing the NI database.}

\item{password}{character. Your password for accessing the NI database.}

\item{eco}{character; optional vector specifying ecosystems to include in the
dataset. Ignored with nonempty \code{indic}.}

\item{indic}{character; optional vector specifying indicators to include in
the dataset.}

\item{year}{character; optional vector specifying years to include in the
dataset.}

\item{norwegian}{logical; use norwegian (\code{TRUE}, default) or english
(\code{FALSE}) names for indicators and ecosystems.}

\item{refYearCode}{integer; code used for reference values. Defaults to
\code{0} which is the code used in the NI database.}
}
\value{
A list of class \code{\link{niDataImport}} with
  the following elements:
  \tabular{lll}{\code{[[1]]} \tab \code{$indicators} \tab a data.frame with indicator data.\cr
     \code{[[2]]} \tab \code{$referenceValues} \tab a list of reference values.\cr
     \code{[[3]]} \tab \code{$indicatorObservations} \tab a list of indicator observations.\cr
     \code{[[4]]} \tab \code{$ICunits} \tab a data.frame describing ICunits.\cr
     \code{[[5]]} \tab \code{$BSunits} \tab a data.frame with BSunit data.\cr
     \code{[[6]]} \tab \code{$ecosystems} \tab a data.frame with ecosystem names and IDs.}
  The vignette \code{objectsInNIcalc} gives a more detailed description of
  \code{niDataImport} lists.
}
\description{
Imports a data set from the Norwegian Nature Index database via API
}
\details{
\code{importDatasetApi} imports a complete data set for calculating the Nature
Index or a thematic index by reading various tables in the NI database via the
[The Nature Index Application Programming Interface](https://www8.nina.no/NaturindeksNiCalc/index.html)
and returns the data set as a list of class \code{\link{niDataImport}}.

The function includes options for assembling complete data sets for an ecosystem,
for e.g. subsequent calculation of the Nature Index. It also includes options
for assembling a data set for a particular set of indicators, for e.g. subsequent
calculation of a thematic index. Further, it includes options for selecting data
from particular years, and for choosing between Norwegian (default) and English
ecosystem and indicator names.

\code{importDatasetApi} contains some simple procedures for data washing.
It removes from the data set
\itemize{
\item passive indicators and indicators with zero fidelities to ecosystems in
question.
\item redundant BSunits and ICunits, i.e. units with no indicator observations
from the ecosystem in question.
\item ICunits without reference values or with reference values equal zero.
\item ICunits with only missing observations or with only reference values
and no other observations.
}
}
\examples{
\dontrun{
# Import data for the calculation of a Nature Index for forests in Norway
# covering the years 1990, 2000, 2010, 2014, 2019.

forestImport <- importDatasetApi(username = "your username",
                                 password = "your password",
                                 eco = "Skog",
                                 indic = NULL,
                                 year = c("1990","2000","2010","2014","2019"),
                                 norwegian = TRUE,
                                 refYearCode = 0)

# Import data for the calculation of a thematic index for game birds in Norway
# in 2000

amphibImport <- importDatasetApi(username = "your username",
                                 password = "your password",
                                 eco = NULL,
                                 indic = c("Fjellrype","Lirype","Orrfugl","Storfugl"),
                                 year = c("2000"),
                                 norwegian = TRUE,
                                 refYearCode = 0)

# Import data for the calculation of an indicator index

cormorImport <- importDatasetApi(username = "your username",
                                 password = "your password",
                                 eco = NULL,
                                 indic = c("Toppskarv"),
                                 year = c("1990","2000","2010","2014","2019"),
                                 norwegian = TRUE,
                                 refYearCode = 0)

}

}
\seealso{
\code{\link{getNIdata}} for functions called by \code{importDatasetApi}
that import data from specific tables in the NI database, and
\code{\link{niDataImport}}.
}
\author{
Bård Pedersen
}
