% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impStand.R
\encoding{UTF-8}
\name{impStand}
\alias{impStand}
\title{Standardize Imputations}
\usage{
impStand(x = NULL, imputations = NULL)
}
\arguments{
\item{x}{list of class \code{niInput}.}

\item{imputations}{list of class \code{niImputations}}
}
\value{
A list of class \code{niImputations}
}
\description{
Optional *post hoc* adjustment of sets of multiple imputations for a set of
missing indicator observations.
}
\details{
In some instances, especially when working with very unbalanced data, the
imputation model used in \code{\link{imputeData}} may not produce neutral
imputations as desired. In such cases, *post hoc* adjustment of imputed
values may be used to avoid that imputations introduce new patterns in the
data set. In this context, neutral values are imputations with an expected
deviation from the (grand) mean equal to the sum of 1) the mean deviation for
observations from the year in question, and 2) the mean deviation from annual
means for observations of the indicator in question.
}
\examples{
\dontrun{
imputedValues <- imputeData(x = themeData,
                            nSim = 1000,
                            transConst = 0.01,
                            maxit = 20,
                            printFlag = TRUE)
imputedValues <- impStand(x = themeData,
                          imputations = imputedValues)
}

}
\seealso{
\code{\link{imputeData}}, \code{\link{imputeDiagnostics}},
\code{\link[mice:mice]{mice::mice}}. \cr The vignette \code{objectsInNIcalc}
gives a detailed description of \code{\link{niImputations}} and
\code{\link{niInput}} lists.
}
\author{
Bård Pedersen
}
