% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeDiagnostics.R
\encoding{UTF-8}
\name{imputeDiagnostics}
\alias{imputeDiagnostics}
\title{Imputation Diagnostics}
\usage{
imputeDiagnostics(x = NULL, nSim = 10, transConst = 0.01, ...)
}
\arguments{
\item{x}{list of class \code{niInput}.}

\item{nSim}{integer, number of imputations, default is 10.}

\item{transConst}{numeric scalar, \code{0<transConst<=0.1}.
Transformation constant in log-transformation. Default is \code{0.01}.}

\item{...}{further arguments passed on to \code{mice::mice}.}
}
\value{
A list of three elements:
  \cr\code{identifiers}: a data.frame with variables relating each imputed
  indicator observation to a missing observation in the data set \code{x}.
  \cr\code{imp}: list with 21 elements containing the complete output
  from \code{mice::mice}.
  \cr\code{diagnostics}: list of diagnostic outputs including convergence
  plots, stripplots showing the distribution of imputed and observed
  values, and plots showing the distribution of missing values over years
  and indicators.
}
\description{
Performs multivariate imputation by chained equations and returns a set of
diagnostic plots and statistics together with the imputations.
}
\details{
\code{imputeDiagnostics} is essentially the same function as
\code{\link{imputeData}}, but returns in addition the complete output from
\code{mice::mice} as well as a series of diagnostic plots.

\code{imputeData} will generally involve lengthy calculations where the
\code{nSim} argument is set equal to the number of bootstrap samples in
subsequent calculation of an index using \code{\link{calculateIndex}}.

\code{imputeDiagnostics} is intended to be run before \code{imputeData},
allowing inspection of the diagnostics using a low value for \code{nSim},
and thus facilitating trial runs with different settings of input arguments
to \code{mice::mice}.
}
\examples{
\dontrun{
imputeDiagnostics(x = themeData,
                 nSim = 10,
                 maxit = 100)$diagnostics
}

}
\seealso{
\code{\link{imputeData}} and \code{\link[mice:mice]{mice::mice}}.
\cr The vignette \code{objectsInNIcalc} gives a detailed description of
\code{\link{niInput}} lists.
}
\author{
Bård Pedersen
}
