% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Indic}
\alias{Indic}
\title{Indic}
\format{
A data frame with 33 rows and 6 variables:
\describe{
  \item{\code{IndicatorID}}{integer COLUMN_DESCRIPTION}
  \item{\code{Indicator_name}}{character COLUMN_DESCRIPTION}
  \item{\code{TrophicgroupID}}{integer COLUMN_DESCRIPTION}
  \item{\code{Scalingmodel}}{integer COLUMN_DESCRIPTION}
  \item{\code{Key.indicators}}{logical COLUMN_DESCRIPTION}
  \item{\code{Fidelity}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Indic
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
