% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indicator_area_ind.R
\docType{data}
\name{Indicator.area.ind}
\alias{Indicator.area.ind}
\title{Example dataset of indicator area IDs}
\format{A data frame with 59 rows and 3 columns:
\describe{
  \item{IndicatorareaID}{Numerical ID of indicators areas}
  \item{IndicatorID}{Numerical ID of indicators}
  \item{Indicator_name}{Character name of indicators}
  }}
\usage{
Indicator.area.ind
}
\description{
A dataset of 59 indicator areas to be used in the examples
}
\keyword{datasets}
