% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Indicator.area}
\alias{Indicator.area}
\title{Indicator.area}
\format{
A data frame with 67 rows and 3 variables:
\describe{
  \item{\code{IndicatorareaID}}{integer COLUMN_DESCRIPTION}
  \item{\code{BasicsunitID}}{integer COLUMN_DESCRIPTION}
  \item{\code{Basicunit}}{integer COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Indicator.area
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
