% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicatorData}
\alias{indicatorData}
\title{Example of an indicatorData set.}
\format{
An object of class \code{indicatorData}  with two elements, a list containing
1) a data frame \code{indicatorData$indicatorValues} with 30 rows and the following
variables
\describe{
  \item{indicatorId}{integer}
  \item{indicatorName}{character}
  \item{areaId}{integer}
  \item{areaName}{character}
  \item{yearId}{integer}
  \item{yearName}{character}
  \item{verdi}{numeric, the expected value of the observation}
  \item{nedre_Kvartil}{numeric, the lower quartile of the observation}
  \item{ovre_Kvartil}{numeric, the upper quartile of the observation}
  \item{datatypeId}{integer}
  \item{datatypeName}{character, type of observation}
  \item{unitOfMeasurement}{character}
  \item{customDistributionUUID}{character, id of corresponding distribution object}
  \item{distributionName}{character, name of fitted distribution}
  \item{distributionId}{integer}
  \item{distParam1}{numeric, first parameter of fitted distribution}
  \item{distParam2}{numeric, second parameter of fitted distribution}
}
2) a list of distribution objects identifying the uncertainties of the estimates (optional)
}
\usage{
indicatorData
}
\description{
A dataset containing observations of the `Dagsommerfugler` indicator
as an example of an object of class \code{indicatorData}.
}
\seealso{
\code{\link{setIndicatorValues}}, and \code{\link{getIndicatorValues}}
}
\keyword{datasets}
