% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCustomUncertainty.R
\name{makeCustomUncertainty}
\alias{makeCustomUncertainty}
\title{Create a distribution function to describe the uncertainty for an indicator}
\usage{
makeCustomUncertainty(input = NULL, functionParams = NULL)
}
\arguments{
\item{input}{Either a name of a known distribution, a vector of values, or a data frame of possibleValues and valueProbabilities}

\item{functionParams}{(optional) Parameters for the distribution function, if such is provided in `input`}
}
\value{
an object of class `NIdistribution`
}
\description{
This functions formats various representations of indicator uncertainty into a common structure for further processing
}
\examples{

 makeCustomUncertainty(dnorm)

}
\seealso{
\code{\link{sample.NIdistribution}}
}
\author{
Jens Åström
}
