% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanTestObject.R
\name{mean.testObject}
\alias{mean.testObject}
\title{Methods for class testObject}
\usage{
\method{mean}{testObject}(x)
}
\description{
I think this is the wrong way to go about it.
}
\details{
Should probably document different methods separately
}

