% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munweights.R
\name{munweights}
\alias{munweights}
\title{Weighting for municipalities}
\usage{
munweights(Municipalities = paste("Mun", 1:36, sep = ""),
  Indicators = c("Ind1", "Ind2", "Ind3", "Ind4", "Ind5", "Ind6"),
  FK_TrophicgroupID = c(1, 1, 2, 2, 0, 0), Key.indicators = c(F, F, F,
  F, T, T), Fidelity = rep(1, 6), Areaind.Name = 1:36,
  Areaind.Indicator = rep(c("Ind1", "Ind2", "Ind3", "Ind4", "Ind5",
  "Ind6"), each = 6), Area.Name = rep(1:36, each = 5),
  Area.Municipality = paste("Mun", 1:36, sep = "")[c(sample(1:36, 30),
  sample(1:36, 30), sample(1:36, 30), sample(1:36, 30), sample(1:36, 30),
  sample(1:36, 30))])
}
\arguments{
\item{Municipalities}{character vector of length=length(Municipality) with names of basic spatial units.}

\item{Indicators}{character    vector of    length=length(Indicators)    with the names of all indicators included in the calculation.}

\item{FK_TrophicgroupID}{integer    vector of    length=length(Indicators) with trophic group IDs that the corresponding indicator belongs to.}

\item{Key.indicators}{logical    vector of    length=length(Indicators), indicating whether the corresponding indicator is a key indicator or not.}

\item{Fidelity}{numeric vector of length=length(Indicators)    with fidelities of each indicator to the major habitat in question.}

\item{Areaind.Name}{integer    vector of    length=length(Areaind.Name)    with unique indicator area IDs.}

\item{Areaind.Indicator}{character    vector of length=length(Areaind.Name)    with indicator names corresponding to the entries in Areaind.Name.}

\item{Area.Name}{integer    vector of length=length(Area.Name)    with indicator area IDs, where each indicator area ID is
repeated for each spatial unit it consist of.}

\item{Area.Municipality}{character        length=length(Area.Name)    vector of names of basic spatial units corresponding to the entries in Area.Name.}
}
\value{
Weights.trof numeric matrix of	dim=length(Municipality) x length(Indicators) Wi * Wf.
}
\description{
Calculates fidelity weights (Wi), trophic weights (Wf) and their product for each combination
of municipality and indicator.
}
\details{
Wi is the fidelity of indicator to ecosystem in questions divided by summed fidelity
of indicators belonging to the same trophic group occuring in a municipality.
Wf for non-key indicators equals 0.5/number of trophic groups observed in the municipality if also
key indicators are observed in the municipality, and 1/number of trophic groups observed in the municipality
if no key elements are observed in the municipality. Wf for key indicators equals 0.5 if also
non-key indicators are observed in the municipality, and 1 if only key indicators are observed.
Weights.trof : Matrix Double (N.kom x N.ind) Wi * Wf
}
\section{Notes}{

}

\examples{
\dontrun{
##Add example, here, possibly as a don't run
tt <- munweights.fct(
  Municipalities=paste("Mun", 1:36, sep=""),
  Indicators=c("Ind1","Ind2","Ind3","Ind4","Ind5","Ind6"),
  FK_TrophicgroupID=c(1,1,2,2,0,0),
  Key.indicators=c(F,F,F,F,T,T),
  Fidelity=rep(1,6),
  Areaind.Name=1:36,
  Areaind.Indicator=rep(c("Ind1","Ind2","Ind3","Ind4","Ind5","Ind6"),each=6),
  Area.Name=rep(1:36,each=5),
Area.Municipality=paste("Mun", 1:36, sep="")[c(sample(1:36,30), sample(1:36,30), sample(1:36,30),
                                                sample(1:36,30), sample(1:36,30), sample(1:36,30))])
}

}
\author{
Bård Pedersen
}
