% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIdata1.R
\docType{data}
\name{NIdata1}
\alias{NIdata1}
\title{Dummy dataset}
\format{A data frame with 40 rows and 5 columns:
\describe{
  \item{column1}{random numbers}
  \item{column1}{random numbers}
  \item{column1}{random numbers}
  \item{column1}{random numbers}
  \item{column1}{random numbers}
  }}
\usage{
NIdata1
}
\description{
A dataset of 200 random numbers put in a dataframe
}
\keyword{datasets}

