% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niImputations.R
\encoding{UTF-8}
\name{niImputations}
\alias{niImputations}
\alias{is.niImputations}
\title{Create or test for Object of Class \code{niImputations}}
\usage{
niImputations(identifiers = NULL, imputations = NULL)

is.niImputations(x)
}
\arguments{
\item{identifiers}{data.frame}

\item{imputations}{numeric matrix of imputed indicator observations}

\item{x}{object to be tested (\code{is.niImputations}).}
}
\value{
\code{niImputations} returns an \code{S3} object, a list of class
   \code{niImputations}.\cr
   \cr\code{is.niImputations} returns \code{TRUE} if the argument is of class
   \code{niImputations}
}
\description{
\code{niImputations} assembles arguments and returns an \code{S3} object of class
\code{niImputations}.
\cr\code{is.niImputations} tests if a candidate object is of class
\code{niImputations}.
}
\details{
Objects of class \code{niImputations} contain imputed values for missing values
in a set of indicator observations.

Class \code{niImputations} objects are generated by the function
\code{imputeData}.
}
\seealso{
Functions \code{\link{imputeData}} and \code{\link{imputeDiagnostics}}.
\cr The vignette \code{objectsInNIcalc} gives a more detailed description of
\code{\link{niImputations}} lists.
}
\author{
Bård Pedersen
}
