% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niInput.R
\encoding{UTF-8}
\name{niInput}
\alias{niInput}
\alias{is.niInput}
\title{Create or test for Object of Class \code{niInput}}
\usage{
niInput(
  indicators = NULL,
  ICunits = NULL,
  BSunits = NULL,
  referenceValues = NULL,
  indicatorValues = NULL,
  NIunits = NULL
)

is.niInput(x)
}
\arguments{
\item{indicators}{data.frame of indicator data.}

\item{ICunits}{integer BSunit x indicator matrix of ICunits.}

\item{BSunits}{data.frame of BSunit data.}

\item{referenceValues}{data.frame of reference values.}

\item{indicatorValues}{list of data.frames of indicator values.}

\item{NIunits}{0/1 BSunit x NIunit matrix of NIunits.}

\item{x}{object to be tested (\code{is.niInput})}
}
\value{
\code{niInput} returns a list of class \code{niInput} with elements:
   \tabular{ll}{\code{[[1]]} \tab \code{indicators}\cr
   \code{[[2]]} \tab \code{ICunits}\cr
   \code{[[3]]} \tab \code{BSunits}\cr
   \code{[[4]]} \tab \code{referenceValues}\cr
   \code{[[5]]} \tab \code{indicatorValues}\cr
   \code{[[6]]} \tab \code{NIunits}}
   \code{is.niInput} returns \code{TRUE} if the argument is of class
   \code{niInput}
}
\description{
\code{niInput} assembles arguments and returns an \code{S3} object of class
\code{niInput}.
\cr\code{is.niInput} tests if a candidate object is of class \code{niInput}.
}
\details{
Lists of class \code{niInput} contain data sets that are controlled for consistency
and for including all necessary data for calculating the Nature Index or a
thematic index. \code{niInput} objects have the particular structure and format
required for data input to the function \code{calculateIndex}. The function
\code{assembleNiObject} returns objects of class \code{niInput}.
}
\seealso{
\code{\link{assembleNiObject}}, \code{\link{calculateIndex}}.
\cr The vignette \code{objectsInNIcalc} gives a more detailed description
of \code{niInput} lists.
}
\author{
Bård Pedersen
}
