% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal2Lognormal.R
\name{normal2Lognormal}
\alias{normal2Lognormal}
\title{Transform normal distribution to log-normal}
\usage{
normal2Lognormal(mean = NULL, sd = NULL)
}
\arguments{
\item{sd}{numerica the year of the values to retrieve. Default = NULL, which means you get all values.}

\item{mu}{numerical    a numerical value identifying the indicator. Usually this is first retreived by the \link{getIndicators} function}

\item{type}{Return transformed mu or standard deviation? Defaults to mu}
}
\value{
A list of "mean" and "sd", giving the mean and sd of the lognormal distribution, suited to be used later in e.g. \code{gamlss.dist}
}
\description{
This is a convenience function that takes the mean and the standard deviation of a normal distribution and returns the corresponding values for the lognormal distibution of the values.
}
\examples{
\dontrun{


lognormalPars <- normal2Lognormal(mean = 100, sd = 10)


 }

}
\author{
Jens Åström, Bård Pedersen
}
