% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Observations}
\alias{Observations}
\title{Observations}
\format{
A data frame with 295 rows and 7 variables:
\describe{
  \item{\code{ValueID}}{integer COLUMN_DESCRIPTION}
  \item{\code{IndicatorID}}{integer COLUMN_DESCRIPTION}
  \item{\code{Expected.value}}{double COLUMN_DESCRIPTION}
  \item{\code{Lower}}{double COLUMN_DESCRIPTION}
  \item{\code{Upper}}{double COLUMN_DESCRIPTION}
  \item{\code{ReferenceYearID}}{integer COLUMN_DESCRIPTION}
  \item{\code{IndicatorareaID}}{integer COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Observations
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
