% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internfunctions.R
\docType{data}
\name{passEnv}
\alias{passEnv}
\title{Internal functions}
\format{
An object of class \code{environment} of length 0.
}
\usage{
passEnv
}
\value{
Builds a passive/hidden environment
}
\description{
These functions set up a passive environment to store tokens
}
\author{
Jens Åström
}
\keyword{internal}
