% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.niSeries.R
\encoding{UTF-8}
\name{plot.niSeries}
\alias{plot.niSeries}
\title{Plot a Time Series of Index Values.}
\usage{
\method{plot}{niSeries}(x = NULL, ...)
}
\arguments{
\item{x}{an \code{niSeries} object.}

\item{...}{further arguments controlling the appearance of the plot,
some passed further to \code{\link{par}}:
\cr\code{type}, \code{linecol}, \code{lwd}, \code{cex}, \code{cex.axis}, \code{main}, \code{cex.main},
\cr\code{gridlines} - whether to include gridlines (\code{T}, default) or not (\code{F}),
\cr\code{gridcol} - grid line colour (defaults to \code{gray(0.6)}),
\cr\code{showInterval} - weather to include interval estimates (\code{T}, default) in the plot.
\cr\code{limits} - limit quantiles for interval estimate (defaults to \code{c(0.025,0.975)})
\cr\code{shade} - weather to include a shading (\code{T}) to represent interval estimate or
   not (\code{F}, default),
\cr\code{polygoncol} color of shade,
\cr\code{whiskerEnds} weather to draw ends of the interval estimate, with (\code{T}, default)
   or without (\code{F}) whisker ends.}
}
\description{
Plot method for objects of class \code{\link{niSeries}}.
}
\examples{
\dontrun{
# Plotting the results from the calculation of a thematic index:

plot(calculateIndex(x = themeData,
                             nsim = 1000,
                             fids = FALSE,
                             tgroups = FALSE,
                             keys = "ignore",
                             w = 0,
                             awbs = TRUE,
                             awBSunit = "Skog")$E,
    cex=1.5,
    whiskerEnds = TRUE)
}

}
\seealso{
\code{\link{calculateIndex}}, \code{\link{niSeries}}, and
\code{\link{niOutput}} for cumputation
and generation of \code{niSeries} and \code{niOutput} objects.
}
\author{
Bård Pedersen
}
