% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.niValue.R
\encoding{UTF-8}
\name{plot.niValue}
\alias{plot.niValue}
\title{Plot the Distribution of an Index Estimate.}
\usage{
\method{plot}{niValue}(x = NULL, ...)
}
\arguments{
\item{x}{\code{niValue} object.}

\item{...}{further arguments controlling the appearance of the plot,
some passed further to \code{\link{par}}:
\cr\code{linecol}, \code{lwd}, \code{xlab}, \code{ylab}, \code{cex.lab}, \code{cex.axis}, \code{main},
   \code{cex.main},
\cr\code{q} - vector of quantiles displayed along first (as \code{quantile(q)}) and second axis (as \code{q})
   (defaults to \code{c(0.025,0.500,0.975)}),
\cr\code{gridcol} - grid line colour (defaults to \code{gray(0.6)}) where gridlines are defined by \code{q}),
\cr\code{range01} - whether the x-axis should cover the full index range (\code{c(0,1)},\code{T}) or
   not (\code{F}).}
}
\description{
Plot method for objects of class \code{\link{niValue}}.
}
\details{
Plots the cumulative distribution of an index value.
}
\examples{
\dontrun{
# Plotting the cumulative distribution function of a thematic index value:

plot(calculateIndex(x = themeData,
                    nsim = 1000,
                    fids = FALSE,
                    tgroups = FALSE,
                    keys = "ignore",
                    w = 0,
                    awbs = TRUE,
                    awBSunit = "Skog")$E$`1990`,
     range01 = F)
}

}
\seealso{
\code{\link{calculateIndex}}, \code{\link{niValue}}, \code{\link{niSeries}},
and \code{\link{niOutput}} for computation and generation of \code{niValue},
\code{niSeries}, and \code{niOutput} objects.
}
\author{
Bård Pedersen
}
