% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\encoding{UTF-8}
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot function for known distributions}
\usage{
plotDistribution(
  distrib = c("Gamma", "ZIExponential", "LogNormal", "TruncNormal", "Weibull"),
  mu = NULL,
  sig = NULL,
  refValue = 1,
  obsval = c(lower = NULL, mu = NULL, upper = NULL),
  type = "continuous"
)
}
\arguments{
\item{distrib}{Type of distribution. Must match allowed values.}

\item{mu}{first parameter of distribution}

\item{sig}{second parameter of distribution}

\item{refValue}{reference value to be plotted along with the distribution}

\item{obsval}{double, length = 3, observed mean and quartiles in the sequence
\code{c(lower.quartile, mean, upper.quartile)}}

\item{type}{fixed parameter set to "continuous"}
}
\description{
Plot function for known distributions, defined by input parameters
}
\note{
In it's present form \code{plotDistribution} covers the Gamma-
lognormal, truncated normal, Weibull and ZIExponential distributions. It may
be expanded to cover more distributions and to accept other parameters than
expected values and quartiles in \code{obsval}.
}
\seealso{
\code{\link{fitAndPlotDistribution}}
}
\author{
Jens Åström, Bård Pedersen
}
