% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeights.R
\encoding{UTF-8}
\name{plotWeights}
\alias{plotWeights}
\title{Barplot showing Weights per Indicator, trophic Group, BSunit or ICunit.}
\usage{
plotWeights(x = NULL, group = "indic", ...)
}
\arguments{
\item{x}{an \code{niValue} object.}

\item{group}{character. Should summary be
\cr per indicator (\code{group = "indic"}, default),
\cr per trophic group (\code{group = "troph"}),
\cr per BSunit (\code{group = "BSunit"}), or
\cr per ICunit (\code{group = "ICunit"})?}

\item{...}{further arguments controlling the appearance of the plot,
some passed further to \code{\link{par}}:
\cr\code{lwd}, \code{cex}, \code{cex.axis}, \code{main}, \code{cex.main},
\cr\code{barnum} - maximum number of bars (default = 35),
\cr\code{barcol} - color for the bars (defaults to \code{gray(0.75)}),
\cr\code{keys} - whether key indicators should be shown in a separate color (\code{T}, default),
\cr\code{keycol} color for key indicator bars.}
}
\description{
Plot method for objects of class \code{\link{niValue}} which creates a horizontal barplot.
}
\examples{
\dontrun{

plotWeights(calculateIndex(x = themeData,
                           nsim = 1000,
                           fids = FALSE,
                           tgroups = FALSE,
                           keys = "ignore",
                           w = 0,
                           awbs = TRUE,
                           awBSunit = "Skog")$E$`1990`,
             group = "ICunit",
             barcol = 3)
}

}
\seealso{
\code{\link{calculateIndex}}, \code{\link{niValue}},
\code{\link{niSeries}}, and \code{\link{niOutput}} for cumputation
and generation of \code{niValue}, \code{niSeries}, and \code{niOutput} objects.
}
\author{
Bård Pedersen
}
