% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printIndicatorData.R
\name{print.indicatorData}
\alias{print.indicatorData}
\alias{print.niIndicatorData}
\title{Print function for class indicatorData objects}
\usage{
print.niIndicatorData(input)
}
\arguments{
\item{input}{object of class `indicatorData`}
}
\value{
Vector of samples
}
\description{
Simple print function that displays also the custom distributions.
}
\examples{
myDist <- makeCustomUncertainty(input = "Normal", distParams = list("mean" = 1, "sd" = 0.2))
sampleNIDistribution(myDist, 10)



}
\seealso{
\code{\link{makeCustomDistribution}}
}
\author{
Jens Åström
}
