% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNIdistribution.R
\name{sample.Distribution}
\alias{sample.Distribution}
\alias{sampleNIDistribution}
\title{Sample function for Nature index distributions}
\usage{
sampleNIDistribution(dist, nSamples = 10)
}
\arguments{
\item{dist}{object of class `NIdistribution` created by \code{makeCustomDistribution}}

\item{nSamples}{Number of samples to draw. Integer.}
}
\value{
Vector of samples
}
\description{
This is a simple convenience function to draw samples from object made from NIdistributions
}
\examples{
myDist <- makeCustomUncertainty(input = "Normal", distParams = list("mean" = 1, "sd" = 0.2))
sampleNIDistribution(myDist, 10)



}
\author{
Jens Åström
}
\seealso{
\code{\link{makeCustomDistribution}}
}

