% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDistribution.R
\name{sampleDistribution}
\alias{sampleDistribution}
\title{Sample function for Nature index distributions}
\usage{
sampleDistribution(dist, nSamples = 10)
}
\arguments{
\item{dist}{object of class `NIdistribution` created by \code{makeDistribution}}

\item{nSamples}{Number of samples to draw. Integer.}
}
\value{
Vector of samples
}
\description{
This is a simple convenience function to draw samples from object made from NIdistributions
}
\examples{
myDist <- makeDistribution(input = "Normal", distParams = list("mean" = 1, "sd" = 0.2))
sampleNIDistribution(myDist, 10)



}
\seealso{
\code{\link{makeDistribution}}
}
\author{
Jens Åström
}
