% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleobs.R
\name{sampleobs}
\alias{sampleobs}
\title{Random Draws}
\usage{
sampleobs(nsim = 99, distrib, mu, sig, obs)
}
\arguments{
\item{nsim}{double    length = 1, number of draws (simulations)}

\item{distrib}{character    length = 1, distribution family ("Normal","TruncNormal","LogNormal","Gumbel","Weibull","ZIExponential","Gamma","Poisson","NegBinom","ZIP") or copy obs (when distrib == "NoBoot")}

\item{mu}{double    length = 1, parameter 1 of model distribution}

\item{sig}{double    length = 1, parameter 2 of model distribution}

\item{obs}{double    length = 1, value (observed mean) to be copied if distrib == "NoBoot".}
}
\value{
double vector length = \code{nsim}, \code{nsim} random draws from
specified (by distrib, mu and sig) model or \code{nsim} copies of \code{obs}.
}
\description{
\code{sampleobs} returns \code{nsim} random draws from a two-parameter
probability distribution specified by the arguments.
}
\details{
\code{distrib} and the parameters mu and sig

from a predetermined set of distribution families. The predetermined set includes
normal-, truncated normal-, lognormal-, weibull-, gumbell-, gamma-, poisson-, negative binomial-, and zero-inflated poisson distribution.
The lower bound in the truncated normal distribution is always zero, while the upper bound is infinity.
With these restrictions all theoretical distributions have two parameters, except the poisson with only one parameter.
Alternatively, if distrib == "NoBoot", the function returns nsim copies of obs.
}
\author{
Nigel Yoccoz and Bård Pedersen
}

