% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleObsMat.R
\encoding{UTF-8}
\name{sampleObsMat}
\alias{sampleObsMat}
\title{Draw Bootstrap Samples}
\usage{
sampleObsMat(
  ICunitId = NULL,
  value = NULL,
  distrib = NULL,
  mu = NULL,
  sig = NULL,
  customDistribution = NULL,
  imputations = NULL,
  obsMethod = "sample",
  obsType = NULL,
  impYear = NULL,
  nsim = 1000
)
}
\arguments{
\item{ICunitId}{integer vector}

\item{value}{numeric vector with expected values}

\item{distrib}{character vector with names of model distributions}

\item{mu}{numeric vector with parameter 1 of model distribution}

\item{sig}{numeric vector with parameter 2 of model distribution}

\item{customDistribution}{list of distribution objects and \code{NA}s}

\item{imputations}{optional list of class \code{niImputations}}

\item{obsMethod}{string, either \code{"sample"} or \code{"constant"}}

\item{obsType}{character vector with observation types. Valid elements are:
\cr\code{"tradObs"}: corresponding observation is a two-parameter,
non-negative, probability distribution given by corresponding elements in
\code{distrib}, \code{mu} and \code{sig}
\cr\code{"customObs"}: corresponding observation is a custom distribution
object in \code{customDistribution}
\cr\code{"imputations"}: corresponding observation is stored in the object
\code{imputations} as \code{nsim} draws, one from each of \code{nsim}
imputed distributions.}

\item{impYear}{character, year in question. The parameter is only relevant
when the data set includes imputed values. Used to identify the correct
elements in \code{imputations} which might include observations from
several years. The parameter is not relevant for sets of reference values.}

\item{nsim}{integer - number of bootstrap simulations, default is 1000.}
}
\value{
numeric matrix, \code{dim = c(length(values), nsim)}, \code{nsim}
bootstrap samples of the set of indicator observations or reference values.
}
\description{
\code{sampleObsMat} returns \code{nsim} bootstrap samples of a set of
indicator observations from the same year or reference values.
}
\details{
Bootstrap samples consist of either \code{nsim} random draws from a set of
probability distributions, or \code{nsim} copies of the elements of
\code{value}, depending on the parameter \code{obsMethod}.

Probability distributions are either two-parameter, non-negative probability
distributions given by vectors \code{distrib}, \code{mu} and \code{sig}, or
distribution objects in the list \code{customDistribution},
or a distribution of \code{nsim} sampled, imputed values in \code{imputes}.

\code{sampleObsMat} presumes that the set of probability distributions only
includes the truncated normal-, lognormal-, weibull-, gamma-, zero-inflated
exponential, poisson-, negative binomial-, and zero-inflated poisson
distributions. The lower bound in the truncated normal distribution is always
zero, while the upper bound is infinity.

\code{sampleObsMat} is a utility function adapted to the mathematical
framework for calculating the nature index.

\code{sampleObsMat} calls functions \code{\link{sampleobs}} and
\code{\link{sampleDistribution}} which do the random draws.
}
\examples{
x <- themeData$indicatorValues$`2019`[11:20,]
sampleObsMat(ICunitId = x$ICunitId,
             value = x$expectedValue,
             distrib = x$distributionFamilyName,
             mu = x$distParameter1,
             sig = x$distParameter2,
             customDistribution = x$customDistribution,
             obsMethod = "sample",
             obsType = "tradObs",
             nsim = 5)

}
\seealso{
\code{\link{sampleobs}}, \code{\link{sampleDistribution}}
}
\author{
Bård Pedersen
}
