% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setIndicatorValues.R
\encoding{UTF-8}
\name{setIndicatorValues}
\alias{setIndicatorValues}
\title{Set indicator values in IndicatorData object}
\usage{
setIndicatorValues(
  indicatorData = NULL,
  areaId = NULL,
  years = NULL,
  est = NA,
  lower = NA,
  upper = NA,
  distribution = NULL,
  datatype = NA,
  unitOfMeasurement = "Enhetsløs"
)
}
\arguments{
\item{indicatorData}{Object of class `indicatorData` created by \code{getIndicatorValues}.}

\item{areaId}{Numeric. Identifier of the indicator value to be altered.}

\item{years}{Numeric. Which year to set values for.}

\item{est}{Numeric. Point estimate for the indicator value. Optional.}

\item{lower}{Numeric. Lower quartile of estimate. Required when supplying point estimate.}

\item{upper}{Numeric. Upper quartile of estimate. Required when supplying point estimate.}

\item{distribution}{distribution object generated by \code{makeDistribution}
See examples therein.}

\item{datatype}{Type of observation. Remember to update this when the type changes.
Allowed values: NA, 1 = Ekspertvurdering, 2 = Overvåkingsdata, 3 = Beregnet fra modeller. Defaults to NA.}

\item{unitOfMeasurement}{Text of maximum length 100. Defaults to "Enhetsløs".}
}
\value{
Object of class `indicatorData`.
}
\description{
This function fills info into an \code{indicatorData} object which can later be uploaded to the database.
Usually, this object is first retrieved from the database through the function \code{getIndicatorData}.
This object will list all data that a user is privileged to alter for a particular indicator.
Data is accepted either in the form of an estimate together with an lower and upper quartile of uncertainty,
or as a distribution object generated by \code{\link{makeDistribution}}.
}
\examples{
myData <- indicatorData
class(myData)
myData <- setIndicatorValues(myData, areaId = 7040, years = 1950,
                             est = 0.9, lower = 0.7, upper = 1,
                             datatype = 1)

myDist <- makeDistribution(input = "logNormal",distParams = list("mean" = 40, "sd" = 2))
myData <- setIndicatorValues(myData, areaId = 7041, years = 1950,
                             distribution = myDist,
                             datatype = 2)

myCodasamples <- rnorm(n = 1000, mean = 15)
myData <- setIndicatorValues(myData, areaId = 7042, years = 1950,
                             distribution = makeDistribution(myCodasamples),
                             datatype = 3, unitOfMeasurement = "antall/m2")


}
\seealso{
\code{\link{makeDistribution}}, and \code{\link{getIndicatorValues}}.
The vignette \code{Distributions} gives detailed descriptions of the use of
\code{setIndicatorValues} when revising and updating values for an indicator.
}
\author{
Jens Åström
}
