% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.niOutput.R
\encoding{UTF-8}
\name{summary.niOutput}
\alias{summary.niOutput}
\title{Summarize Index Value Estimates for a Set of NIunits.}
\usage{
\method{summary}{niOutput}(object = NULL, ...)
}
\arguments{
\item{object}{\code{niOutput} object.}

\item{...}{arguments passed on to \code{\link{summary.niSeries}} controlling
the summary: \cr\code{limits}: quantiles to be used as limits in interval
estimates ("confidence interval", defaults to \code{c(0.025,0.975)}),
\cr\code{locationMean}: which location parameter to use as point estimate
mean (\code{T}) or median (\code{F}, default).}
}
\value{
a list of numeric matrices, one matrix for each NIunit, and each
matrix containing interval and point estimates for each year, together with
an estimate of the displacement parameter (see \code{\link{summary.niSeries}}).
}
\description{
Summary method for objects of class \code{\link{niOutput}}.
}
\examples{
\dontrun{
# Summary of the results from the calculation of a thematic index:

summary(calculateIndex(x = themeData,
                       nsim = 1000,
                       fids = FALSE,
                       tgroups = FALSE,
                       keys = "ignore",
                       w = 0,
                       awbs = TRUE,
                       awBSunit = "Skog"))
}

}
\seealso{
\code{\link{calculateIndex}} and \code{\link{niOutput}} for cumputation
and generation of \code{niOutput} objects.
}
\author{
Bård Pedersen
}
