% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.niSeries.R
\encoding{UTF-8}
\name{summary.niSeries}
\alias{summary.niSeries}
\title{Summarize a Time Series of Index Value Estimates.}
\usage{
\method{summary}{niSeries}(object = NULL, ...)
}
\arguments{
\item{object}{\code{niSeries} object.}

\item{...}{arguments controlling the summary:
\cr\code{limits}: quantiles to be used as limits in intervall
estimates ("confidence interval", defaults to \code{c(0.025,0.975)}),
\cr\code{locationMean}: which location
parameter to use as point estimate mean (\code{T}) or median (\code{F}, default).}
}
\value{
a numeric matrix with interval and point estimates for each year,
together with an estimate of the displacement in the location parameter due to
nonlinear scaling of uncertain indicator observations etc.,
[cf. Pedersen and Skarpaas (2012)](https://www.nina.no/archive/nina/PppBasePdf/rapport/2012/797.pdf).
}
\description{
Summary method for objects of class \code{\link{niSeries}}.
}
\examples{
\dontrun{
# Summary of the results from the calculation of a thematic index:

summary(calculateIndex(x = themeData,
                       nsim = 1000,
                       fids = FALSE,
                       tgroups = FALSE,
                       keys = "ignore",
                       w = 0,
                       awbs = TRUE,
                       awBSunit = "Skog")$E)
}

}
\seealso{
\code{\link{calculateIndex}}, \code{\link{niSeries}}, and
\code{\link{niOutput}} for cumputation and generation of \code{niSeries} and
\code{niOutput} objects.
}
\author{
Bård Pedersen
}
