% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryWeights.R
\encoding{UTF-8}
\name{summaryWeights}
\alias{summaryWeights}
\title{Weights per Indicator, trophic Group, BSunit or ICunit.}
\usage{
summaryWeights(x = NULL, group = "indic", ...)
}
\arguments{
\item{x}{an \code{niSeries} object.}

\item{group}{character. Should summary be per indicator (\code{group = "indic"}, default), per trophic group
(\code{group = "troph"}), per BSunit (\code{group = "BSunit"}), or per ICunit (\code{group = "ICunit"})?}

\item{...}{argument controlling the summary: \code{indicData} logical flag, whether to include additional
variables characterizing indicators (i.e. trophic group, fidelity) in the summary per indicator or not,
default is \code{F}.}
}
\value{
a numeric matrix (when \code{group in c("troph","BSunit") or indicData = F})
or a dataframe (when \code{group = "ICunit" or indicData = T}) with weights
per indicator (\code{group = "indic"}),
per trophic group (\code{group = "troph"}),
per BSunit (\code{group = "BSunit"}), or
per ICunit (\code{group = "ICunit"}),
for each year included in \code{x}.
}
\description{
Summary method for objects of class \code{\link{niSeries}}.
}
\examples{
\dontrun{
indices <- calculateIndex(x = themeData,
                       nsim = 1000,
                       fids = FALSE,
                       tgroups = FALSE,
                       keys = "ignore",
                       w = 0,
                       awbs = TRUE,
                       awBSunit = "Skog")
summaryWeights(indices$wholeArea, indicData = TRUE)
summaryWeights(indices$E, group = "ICunit")
summaryWeights(indices$E, group = "BSunit")
}

}
\seealso{
\code{\link{calculateIndex}}, \code{\link{niSeries}}, and
\code{\link{niOutput}} for cumputation and generation of \code{niSeries} and
\code{niOutput} objects.
}
\author{
Bård Pedersen
}
