% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{themeData}
\alias{themeData}
\title{Example of a thematic index data set.}
\format{
An object of class \code{niInput}  with the following six elements
\describe{
  \item{indicators}{data.frame with indicator data}
  \item{ICunits}{integer BSunit x indicator matrix of ICunits}
  \item{BSunits}{data.frame of BSunit data}
  \item{referenceValues}{data.frame of reference values}
  \item{indicatorValues}{list of data.frames of indicator observations}
  \item{NIunits}{0/1 BSunit x NIunit matrix of NIunits}
}
}
\usage{
themeData
}
\description{
A constructed data set covering four "carnivore" indicators
as an example of data sets of class \code{niInput}.
}
\seealso{
\code{\link{niInput}}. \cr The vignette \code{objectsInNIcalc} gives a
detailed description of \code{\link{niInput}} lists.
}
\keyword{datasets}
