% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeIndicatorValues.R
\name{writeIndicatorValues}
\alias{writeIndicatorValues}
\title{Write indicator values}
\usage{
writeIndicatorValues(indicatorData = NULL)
}
\arguments{
\item{indicatorData}{Data frame of indicator values. Usually format retrieved
from \code{getIndicatorValues} and updated through \code{setIndicatorValues}.}
}
\value{
If posting the data fails, an error message is returned.
}
\description{
This function writes updated values for a given indicator to the Nature Index
database via the Nature Index API.
}
\details{
Functions \code{\link{getIndicatorValues}}, \code{\link{setIndicatorValues}},
and \code{writeIndicatorValues} are used in series to update the set of
values for an indicator. After retrieving the current set of values from the
database using \code{getIndicatorValues}, and updating the set with
\code{setIndicatorValues}, \code{writeIndicatorValues} may be used to post the
updated set to the database.
}
\examples{
\dontrun{
getToken()
myIndicators <- getIndicators()
myValues <- getIndicatorValues(indicatorID = 359)
updatedValues <- setIndicatorValues(myValues, year = 2019, est = 0.3)
writeIndicatorValues(updatedValues)
}
}
\seealso{
\code{\link{getIndicatorValues}}, \code{\link{setIndicatorValues}}
}
\author{
Jens Åström
}
