% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSunits.R, R/Indic.R
\docType{data}
\name{BSunits}
\alias{BSunits}
\title{Example dataset of basic spatial units}
\format{A data frame with 4 rows and 5 columns:
\describe{
  \item{BasicunitID}{Numerical ID of units}
  \item{Basicunit}{Character names of units}
  \item{Area}{Area of units in square meters}
  \item{NIarea1}{Character, indicating the top most level in the spatial hierarchy of the areal units}
  \item{NIarea2}{Character, indicating the second level in the spatial hierarchy of the areal units}
  }}
\usage{
BSunits

BSunits
}
\description{
A dataset of four basic spatial units to be used in the examples

A dataset of 33 indicators to be used in the example calculations
}
\keyword{datasets}

