% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndPlotDistribution.R
\encoding{UTF-8}
\name{fitAndPlotDistribution}
\alias{fitAndPlotDistribution}
\title{Fit and plot distribution functions}
\usage{
fitAndPlotDistribution(
  obsval_mu = NULL,
  obsval_lower = NULL,
  obsval_upper = NULL,
  refValue = NULL,
  probQuant_lower = 0.25,
  probQuant_upper = 0.75,
  type = "continuous",
  ...
)
}
\arguments{
\item{obsval_mu}{input expected value}

\item{obsval_lower}{input lower quartile}

\item{obsval_upper}{input upper quartile}

\item{refValue}{the reference value to be drawn in the  plot}

\item{probQuant_lower}{fixed parameter set to 0.25}

\item{probQuant_upper}{fixed parameter set to 0.75}

\item{type}{fixed parameter set to "continuous"}

\item{...}{further arguments passed on to \code{plotDistribution}}
}
\description{
Calls elicitation functions with inputs and plots the resulting distribution.
}
\details{
\code{fitAndPlotDistribution} calls \code{estim} which fits a two-parameter
non-negative continuous model distribution to an expected value and lower
and upper quartiles provided as input. It uses the least square criterion to
find the best fit among a predetermined set of model distribution families:
\code{c("Gamma","LogNormal", "TruncNormal", "Weibull", "ZIExponential")}.
Thereafter the function calls \code{\link{plotDistribution}} which plot the
distribution together with a reference value.
}
\note{
In it's present form \code{fitAndPlotDistribution} only accepts expected
value and lower and upper quartiles as input to \code{obsval}. It may
be expanded to cover other observed parameters and more distributions.
}
\seealso{
\code{\link{estim}} and \code{\link{plotDistribution}}
}
\author{
Jens Åström
}
