% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndPlotDistribution.R
\name{fitAndPlotDistribution}
\alias{fitAndPlotDistribution}
\title{fitAndPlotDistribution}
\usage{
fitAndPlotDistribution(obsval_mu = 2, obsval_lower = -2, obsval_upper = 5,
  refValue = 4, probQuant_lower = 0.25, probQuant_upper = 0.75,
  type = "continuous", ...)
}
\arguments{
\item{refValue}{the reference value to be drawn in the  plot}

\item{obsval}{named vector containing mu and lower and upper quantiles to fit the distribution to.}

\item{probQuant}{named vector containing the lower and upper quantiles to draw in the plot}
}
\description{
Calls estimate (elicitation) functions with inputs and plots the resulting distribution
}
