% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndicators.R
\name{getIndicators}
\alias{getIndicators}
\title{List the indicators that the user is permitted to alter}
\usage{
getIndicators()
}
\arguments{
\item{token}{character        A character string containing the database API token. Typically created by the function \link{getToken}.}
}
\value{
A data frame containing indicator IDs and names. The indicator names is used in other database functions.
}
\description{
Based on the token in `niToken`, this function lists the indicators (ID and names) that the user
is responsible for
}
\section{Notes}{

}

\examples{
\dontrun{
getIndicators()
}

}
\author{
Jens Åström
}
