% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndicatorValues.R
\encoding{UTF-8}
\name{getIndicatorValues}
\alias{getIndicatorValues}
\title{Get indicator values from the Nature Index database}
\usage{
getIndicatorValues(indicatorID = NULL, years = NULL)
}
\arguments{
\item{indicatorID}{integer    identifying the indicator.
Usually this is first retrieved by the \code{\link{getIndicators}} function}

\item{years}{integer,    the year of the values to retrieve. Default = NULL, which means you get all values.}
}
\value{
Object of class `indicatorData`, A list containing
1) a data frame of the values of the indicator and
2) a list of distribution objects identifying the uncertainties of the estimates (optional)
}
\description{
This function retrieves the current values for a given indicator. It is used both to check the current
information, and as a starting point for updating the data. In case of updates, individual rows are altered
by the function \code{setIndicatorValue}.
}
\examples{
\dontrun{
myValues <- getIndicatorValues(IndicatorID = 1)
}

}
\seealso{
\code{\link{setIndicatorValues}} for updating objects of class `indicatorData`,
and \code{\link{getIndicators}} for list of accessible indicators.
The vignette \code{Distributions} describes the use of these and other functions for
updating data sets.
}
\author{
Jens Åström
}
