% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{Get a connection Token from the NI database}
\usage{
getToken(username = NULL, password = NULL)
}
\arguments{
\item{username}{character        Your username}

\item{password}{character        Your password}
}
\value{
The function sets the variable niToken, which is used in other database functions
}
\description{
This function tries to connect to the NI database API and retreives a token that is later used to
communicate with the database.
}
\section{Notes}{

}
\examples{
\dontrun{
getToken("test.user", "test.password")
}

}
\author{
Jens Åström
}

