% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indic.R
\docType{data}
\name{Indic}
\alias{Indic}
\title{Example dataset of indicator values}
\format{A data frame with 33 rows and 6 columns:
\describe{
  \item{IndicatorID}{Numerical ID of indicators}
  \item{Indicator_name}{Character names of indicators}
  \item{TrophicgroupID}{Numerical ID of trophic group}
  \item{Scalingmodel}{Numerical code, indicating type of scaling model}
  \item{Key.indicators}{Logical, indicating if indicators is key indicators}
  \item{Fidelity}{Numerical values of indicator fidelities in the major ecosystem in the example}
  }}
\usage{
Indic
}
\description{
A dataset of 33 indicators to be used in the example calculations
}
\keyword{datasets}
