% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Indicator.area.ind}
\alias{Indicator.area.ind}
\title{Indicator.area.ind}
\format{
A data frame with 59 rows and 3 variables:
\describe{
  \item{\code{IndicatorareaID}}{integer COLUMN_DESCRIPTION}
  \item{\code{IndicatorID}}{integer COLUMN_DESCRIPTION}
  \item{\code{Indicator_name}}{integer COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Indicator.area.ind
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
