% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indicator_area.R
\docType{data}
\name{Indicator.area}
\alias{Indicator.area}
\title{Example dataset of indicator areas}
\format{A data frame with 67 rows and 3 columns:
\describe{
  \item{IndicatorareaID}{Numerical ID of indicators areas}
  \item{BasicunitID}{Numerical ID of basic spatial units}
  \item{Basicunit}{Character name of basic unit}
  }}
\usage{
Indicator.area
}
\description{
A dataset of 67 indicator areas to be used in the example calculations
}
\keyword{datasets}
