% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDistribution.R
\name{makeDistribution}
\alias{makeDistribution}
\title{Create a distribution function to describe the uncertainty for an indicator}
\usage{
makeDistribution(input = NULL, distParams = NULL)
}
\arguments{
\item{input}{Either a name of a known distribution, a vector of values, or a data frame of possibleValues and valueProbabilities}

\item{distParams}{(optional) Parameters for the distribution function, if such is provided in `input`}
}
\value{
an object of class `NIdistribution`
}
\description{
This functions formats various representations of indicator uncertainty into a common structure for further processing
}
\examples{
myDist <- makeCustomUncertainty(input = "Normal", distParams = list("mean" = 1, "sd" = 0.2))
sampleNIDistribution(myDist, 10)

myDist <- makeCustomUncertainty(input = "Poisson", distParams = list("lambda" = 3))
sampleNIDistribution(myDist, 10)

myProbs <- data.frame("est" = c(0.2, 0.23, 0.34, 0.4), "probs" = c(0.1, 0.4, 0.4, 0.1 ))
myDist <- makeCustomUncertainty(myProbs)
sampleNIDistribution(myDist, 10)

codaSamples <- rnorm(1000, mean = 0.87, sd = 0.3)
myDist <- makeCustomUncertainty(codaSamples)
sampleNIDistribution(myDist, 10)


}
\author{
Jens Åström
}
\seealso{
\code{\link{sample.NIdistribution}}
}

