% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal2Lognormal.R
\name{normal2Lognormal}
\alias{normal2Lognormal}
\alias{logNormal2normal}
\title{Transform normal distribution to log-normal, and vice versa}
\usage{
normal2Lognormal(mean = NULL, sd = NULL)

logNormal2normal(mean = NULL, sd = NULL)
}
\arguments{
\item{sd}{Numerical. Standard deviation (or scale) parameter.}

\item{mu}{Numerical. Mean (or meanlog) parameter.}
}
\value{
A list of "mean" and "sd", giving the mean and scale parameter of the lognormal distribution, suited to be used later in e.g. \code{gamlss.dist}, or the mean and the standard deviation of a the corresponding normal distribution.
}
\description{
This is convenience functions that takes the mean and the standard deviation of a normal distribution and returns the corresponding values for the lognormal distibution of the values, and vice versa.
}
\examples{
\dontrun{


logNormPar <- normal2Lognormal(mean = 100, sd = 10)
logNormal2normal(logNormPar[[1]], logNormPar[[2]])

 }

}
\author{
Jens Åström, Bård Pedersen
}
