% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal2Lognormal.R
\encoding{UTF-8}
\name{normal2Lognormal}
\alias{normal2Lognormal}
\alias{logNormal2normal}
\title{Transform normal distribution to log-normal, and vice versa}
\usage{
normal2Lognormal(mean = NULL, sd = NULL)

logNormal2normal(mean = NULL, sd = NULL)
}
\arguments{
\item{mean}{Numerical. Mean (or meanlog) parameter.}

\item{sd}{Numerical. Standard deviation (or scale) parameter.}
}
\value{
\code{normal2Lognormal}: A list of "mean" and "sd", giving the mean and scale parameter of the
lognormal distribution, suited to be used later in e.g. \code{gamlss.dist}, \cr
\code{logNormal2normal}: A list of mean and standard deviation of a the corresponding normal distribution.
}
\description{
These are two convenience functions that takes the mean and the standard deviation
of a normal distribution and returns the corresponding values for the lognormal
distibution of the values, and vice versa.
}
\examples{
logNormPar <- normal2Lognormal(mean = 100, sd = 10)
logNormal2normal(logNormPar[[1]], logNormPar[[2]])

}
\author{
Jens Åström, Bård Pedersen
}
