% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observations.R
\docType{data}
\name{Observations}
\alias{Observations}
\title{Example dataset of observations}
\format{A data frame with 295 rows and 7 columns:
\describe{
  \item{ValueID}{Numerical ID of indicator values}
  \item{IndicatorID}{Numerical ID of indicators}
  \item{Expexted.value}{Reference value for the indicator}
  \item{Lower}{Numerical value of lower quantile}
  \item{Higher}{Numerical value of upper quantile}
  \item{ReferenceYearID}{Numerical ID of reference Year}
  \item{IndicatorareaID}{Numerical ID of indicators areas}
  }}
\usage{
Observations
}
\description{
A dataset of 259 observations to be used in the examples
}
\keyword{datasets}

