% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution.R
\name{plotDistribution}
\alias{plotDistribution}
\title{plotDistribution}
\usage{
plotDistribution(distrib = c("Gamma", "ZIExponential", "LogNormal",
  "TruncNormal", "Normal", "Weibull", "Poisson", "NegBinom", "ZIP", "NoBoot"),
  mu = NULL, sig = NULL, refValue = 1, obsval = c(lower = NULL, mu =
  NULL, upper = NULL), type = "continuous")
}
\arguments{
\item{distrib}{Type of distribution. Must match allowed values.}

\item{mu}{mu parameter of distribution (i.e. mean for normal distribution)}

\item{sig}{sigma parameter of distribution}

\item{refValue}{reference value to be plotted along with the distribution}

\item{obsval}{}
}
\description{
Plot function for known distributions, defined by input parameters
}
