% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDistribution.R
\encoding{UTF-8}
\name{sampleDistribution}
\alias{sampleDistribution}
\title{Sample distribution functions}
\usage{
sampleDistribution(dist, nSamples = 10)
}
\arguments{
\item{dist}{distribution object to sample}

\item{nSamples}{`integer` size of sample}
}
\value{
a numeric vector containing the sample
}
\description{
Generates random samples from distribution functions/objects
}
\details{
\code{sampleDistribution} can be used for any distribution object generated by
\code{\link{makeDistribution}} or by functions in the \code{distr} package.
It is a wrapper around \code{\link[distr:r]{dist::r}}.
}
\examples{
myDist <- makeDistribution(input = "logNormal", distParams = list("mean" = 1, "sd" = 0.2))
sampleDistribution(dist = myDist, nSamples = 10)

myDist <- makeDistribution(input = "Poisson", distParams = list("lambda" = 3))
sampleDistribution(myDist, 10)

myProbs <- data.frame("est" = c(0.2, 0.23, 0.34, 0.4), "probs" = c(0.1, 0.4, 0.4, 0.1 ))
myDist <- makeDistribution(myProbs)
sampleDistribution(myDist, 10)

codaSamples <- rnorm(10000, mean = 0.87, sd = 0.3)
myDist <- makeDistribution(codaSamples)
mean(sampleDistribution(myDist, 10000))

}
\seealso{
\code{\link{makeDistribution}}
}
\author{
Jens Åström
}
