% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleobs.R
\name{sampleobs}
\alias{sampleobs}
\title{Random Draws}
\usage{
sampleobs(nsim = 99, distrib, mu, sig, obs)
}
\arguments{
\item{nsim}{double    length = 1, number of draws (simulations)}

\item{distrib}{character    length = 1, distribution family ("Normal","TruncNormal","LogNormal","Gumbel","Weibull","ZIExponential","Gamma","Poisson","NegBinom","ZIP") or copy obs (when distrib == "NoBoot")}

\item{mu}{double    length = 1, parameter 1 of model distribution}

\item{sig}{double    length = 1, parameter 2 of model distribution}

\item{obs}{double    length = 1, value (observed mean) to be copied if distrib == "NoBoot".}
}
\value{
double vector length = \code{nsim}, \code{nsim} random draws from
specified (by distrib, mu and sig) model or \code{nsim} copies of \code{obs}.
}
\description{
\code{sampleobs} either returns \code{nsim} random draws from a two-parameter
probability distribution specified by the function's arguments, or \code{nsim}
copies of the argument \code{obs}, depending on the argument \code{distrib}.
}
\details{
\code{sampleobs} samples from a single distribution among a predetermined set
of distribution families. The predetermined set consists of the truncated
normal-, lognormal-, Weibull-, gamma-, zero-inflated exponential-, Poisson-,
negative binomial-, and zero-inflated Poisson distribution families. The lower
bound in the truncated normal distribution is always zero, while the upper
bound is infinity. With these restrictions all distributions in the set have
two parameters, except the Poisson with only one parameter.
Alternatively, if \code{distrib == "NoBoot"}, the function returns \code{nsim}
copies of \code{obs}.

Functions called:
rNO, rtnorm, rLOGNO, rWEI, rGU, qGA, rPO, rNBII, rZIP
random generation functions in the gamlss package
}
\author{
Nigel Yoccoz and Bård Pedersen
}

