% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setIndicatorValues.R
\name{setIndicatorValues}
\alias{setIndicatorValues}
\title{Set indicator values in IndicatorData object}
\usage{
setIndicatorValues(indicatorData = NULL, areaId = NULL, years = NULL,
  est = NULL, lower = NA, upper = NA, distribution = NULL,
  distParams = NULL, datatype = 1, unitOfMeasurement = "Enhetsløs")
}
\arguments{
\item{indicatorData}{Object of class `indicatorData` created by \code{getIndicatorData}.}

\item{years}{Numeric. Which year to set values for.}

\item{est}{Numeric. Point estimate for the indicator. Optional.}

\item{lower}{Numeric. Lower quartile of estimate. Required when supplying point estimate.}

\item{upper}{Numeric. Upper quartile of estimate. Required when supplying point estimate.}

\item{distribution}{Either "logNormal", "Poisson", a vector of values, or a data frame of possible values and value probabilities, to be passed to \code{makeDistribution}. See examples therein.}

\item{datatype}{Type of indicator. Remember to update this when the type changes. Allowed values: 1 = Ekspertvurdering, 2 = Overvåkingsdata, 3 = Beregnet fra modeller. Defaults to 1.}

\item{unitOfMeasurement}{Text of maximum length 100. Defaults to "Enhetsløs".}

\item{areaID}{Numeric. Identifier of the indicator to be altered.}

\item{distrParams}{Distribution parameters to be passed to \code{makeDistribution}. Required when
using a named distribution. See examples in \code{makeDistribution}}
}
\value{
Object of class `indicatorData`.
}
\description{
This function fills info into an indicatorData object which can later be uploaded to the database.
Usually, this object is first retreived from the database through the function \code{getIndicatorData}.
This object will list all indicators that a user is privileged to alter. Data for a particular indicator
is accepted either in the form of an estimate together with an lower and upper quartile of uncertainty,
or as an input to the function \code{makeDistribution}.
}
\details{
In this case, data can be supplied either in the form of a vector of possible values (e.g. posterior samples),
a matrix or dataframe of possible values together with probabilities for each value, or as a named
known distribution. In case of named known distributions, parameter for these distributions is supplied
in distParams as a named list. See \link{makeDistribution} for further details.
}
\examples{
getToken("testUser", "testPassword")
indicatorData <- getIndicators()
setIndicatorValues(indicatorID = 1,
est = 0.5,
lower = 0.2
upper = 0.7
)

setIndicatorValues(indicatorID = 2
distribution = "Normal",
distParams = list("mean" = 0.5,
                  "sd" = 0.1)
)


}
\seealso{
\code{\link{makeDistribution}}
}
\author{
Jens Åström
}
