% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setIndicator.R, R/setIndicatorValues.R
\name{setIndicatorValues}
\alias{setIndicatorValues}
\title{Set indicator values in IndicatorData object}
\usage{
setIndicatorValues(indicatorData = NULL, areaId = NULL, year = NULL,
  est = NULL, lower = NULL, upper = NULL, distribution = NULL,
  distParams = NULL)

setIndicatorValues(indicatorData = NULL, areaId = NULL, year = NULL,
  est = NULL, lower = NULL, upper = NULL, distribution = NULL,
  distParams = NULL)
}
\arguments{
\item{indicatorData}{object of class `indicatorData` created by \code{getIndicatorData}}

\item{est}{Point estimate for the indicator. Optional.}

\item{lower}{lower quartile of estimate. Required when supplying point estimate.}

\item{upper}{upper quartile of estimate. Required when supplying point estimate.}

\item{distribution}{distribution type or empirical values to be passed to \code{makeCustomUncertainty}.}

\item{indicatorID}{Identifier of the indicator to be altered}

\item{distrParams}{distribution parameters to be passed to \code{makeCustomUncertainty}. Required then
using a named distribution}

\item{indicatorData}{object of class `indicatorData` created by \code{getIndicatorData}}

\item{indicatorID}{Identifier of the indicator to be altered}

\item{est}{Point estimate for the indicator. Optional.}

\item{lower}{lower quartile of estimate. Required when supplying point estimate.}

\item{upper}{upper quartile of estimate. Required when supplying point estimate.}

\item{distribution}{distribution type or empirical values to be passed to \code{makeCustomUncertainty}.}

\item{distrParams}{distribution parameters to be passed to \code{makeCustomUncertainty}. Required then
using a named distribution}
}
\value{
Vector of samples

Vector of samples
}
\description{
This function fills info into an indicatorData object which can later be uploaded to the database.
Usually, this object is first retreived from the database through the function \code{getIndicatorData}.
This object will list all indicators that a user is privileged to alter. Data for a particular indicator
is accepted either in the form of an estimate together with an lower and upper quartile of uncertainty,
or as an input to the function \code{makeCustomUncertainty}.

This function fills info into an indicatorData object which can later be uploaded to the database.
Usually, this object is first retreived from the database through the function \code{getIndicatorData}.
This object will list all indicators that a user is privileged to alter. Data for a particular indicator
is accepted either in the form of an estimate together with an lower and upper quartile of uncertainty,
or as an input to the function \code{makeCustomUncertainty}.
}
\details{
In this case, data can be supplied either in the form of a vector of possible values (e.g. posterior samples),
a matrix or dataframe of possible values together with probabilities for each value, or as a named
known distribution. In case of named known distributions, parameter for these distributions is supplied
in distParams as a named list. See \link{makeCustomUncertainty} for further details.

In this case, data can be supplied either in the form of a vector of possible values (e.g. posterior samples),
a matrix or dataframe of possible values together with probabilities for each value, or as a named
known distribution. In case of named known distributions, parameter for these distributions is supplied
in distParams as a named list. See \link{makeCustomUncertainty} for further details.
}
\examples{
indicatorData <- getIndicators()
setIndicator(indicatorID = 1,
est = 0.5,
lower = 0.2
upper = 0.7
)

setIndicators(indicatorID = 2
distribution = "Normal",
distParams = list("mean" = 0.5,
                  "sd" = 0.1)
)


getToken("testUser", "testPassword")
indicatorData <- getIndicators()
setIndicator(indicatorID = 1,
est = 0.5,
lower = 0.2
upper = 0.7
)

setIndicators(indicatorID = 2
distribution = "Normal",
distParams = list("mean" = 0.5,
                  "sd" = 0.1)
)


}
\author{
Jens Åström

Jens Åström
}
\seealso{
\code{\link{makeCustomDistribution}}

\code{\link{makeDistribution}}
}

