% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeIndicatorValues.R
\name{writeIndicatorValues}
\alias{writeIndicatorValues}
\title{Get the present indicator values from the Nature Index database}
\usage{
writeIndicatorValues(indicatorData = NULL, token = niToken)
}
\arguments{
\item{indicatorData}{Data frame of indicator values. Usually format retrieved from \link{getIndicatorValues} and updated through \link{setIndicatorValues}.}

\item{token}{A connection token aquired by \link{getToken}. Simply leave as is if you use `getToken`}
}
\value{
A message from the database API of the turn out of the action.
}
\description{
This function retreives the current values for a given indicator. It is used both to check the current
information, and as a starting point for updating the data. In case of updates, individual rows are altered
by the function setIndicatorValue.
}
\section{Notes}{

}
\examples{
\dontrun{
getToken()
myIndicators <- getIndicators()
myValues <- getIndicatorValues(IndicatorID = 1)
updatedValues <- setIndicatorValues(myValues, year = 2018, est = 0.3)
writeIndicatorValues(updatedValues)
}

}
\author{
Jens Åström
}

